do no double calculation for the timeout

+++ src/AsyncMqttClient.cpp
@@ -388,11 +388,11 @@ void AsyncMqttClient::_onPoll(AsyncClient* client) {
     disconnect();
     return;
   // send ping to ensure the server will receive at least one message inside keepalive window
-  } else if (_lastPingRequestTime == 0 && (millis() - _lastClientActivity) >= (_keepAlive * 1000 * 0.7)) {
+  } else if (_lastPingRequestTime == 0 && (millis() - _lastClientActivity) >= (_keepAlive * static_cast<uint32_t >(1000 * 0.7))) {
     _sendPing();
 
   // send ping to verify if the server is still there (ensure this is not a half connection)
-  } else if (_connected && _lastPingRequestTime == 0 && (millis() - _lastServerActivity) >= (_keepAlive * 1000 * 0.7)) {
+  } else if (_connected && _lastPingRequestTime == 0 && (millis() - _lastServerActivity) >= (_keepAlive * static_cast<uint32_t >(1000 * 0.7))) {
     _sendPing();
   }
